back_wpp/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib